#1674 added `CUSTOM` enum value to ContainerSize. ContainerSize moved to subutai-hub-share library.
