+++ management/server/core/environment-manager/environment-manager-cli/src/main/java/io/subutai/core/environment/cli/BuildLocalEnvironmentCommand.java
@@ -11,7 +11,7 @@ import com.google.common.base.Preconditions;
 
 import io.subutai.common.environment.Node;
 import io.subutai.common.environment.Topology;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.common.peer.ResourceHost;
 import io.subutai.core.environment.api.EnvironmentManager;
 import io.subutai.core.identity.rbac.cli.SubutaiShellCommandSupport;
