+++ management/server/core/environment-manager/environment-manager-impl/src/main/java/io/subutai/core/environment/impl/workflow/modification/steps/ChangeQuotaStep.java
@@ -3,7 +3,7 @@ package io.subutai.core.environment.impl.workflow.modification.steps;
 
 import java.util.Map;
 
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.common.peer.EnvironmentContainerHost;
 import io.subutai.common.peer.PeerException;
 import io.subutai.common.tracker.TrackerOperation;
