+++ management/server/core/lxc-quota-manager/lxc-quota-manager-impl/src/main/java/io/subutai/core/lxc/quota/impl/QuotaManagerImpl.java
@@ -33,7 +33,7 @@ import io.subutai.common.metric.ResourceHostMetric;
 import io.subutai.common.metric.ResourceHostMetrics;
 import io.subutai.common.peer.ContainerHost;
 import io.subutai.common.peer.ContainerId;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.common.peer.HostNotFoundException;
 import io.subutai.common.peer.LocalPeer;
 import io.subutai.common.peer.PeerException;
@@ -123,6 +123,10 @@ public class QuotaManagerImpl implements QuotaManager
         int i = 0;
         for ( ContainerSize containerSize : ContainerSize.values() )
         {
+            if ( containerSize == ContainerSize.CUSTOM || i >= settings.length )
+            {
+                break;
+            }
             LOGGER.debug( String.format( "Settings for %s: %s", containerSize, settings[i] ) );
             String[] quotas = settings[i++].split( "\\|" );
