+++ management/server/core/strategy-manager/strategy-manager-impl/src/main/java/io/subutai/core/strategy/impl/AllocatedContainer.java
@@ -1,7 +1,7 @@
 package io.subutai.core.strategy.impl;
 
 
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 
 
 /**
