+++ management/server/core/strategy-manager/strategy-manager-impl/src/main/java/io/subutai/core/strategy/impl/ExamplePlacementStrategy.java
@@ -14,7 +14,7 @@ import org.slf4j.LoggerFactory;
 import io.subutai.common.environment.Node;
 import io.subutai.common.environment.NodeSchema;
 import io.subutai.common.environment.Topology;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.core.strategy.api.ExampleStrategy;
 import io.subutai.core.strategy.api.StrategyException;
 import io.subutai.hub.share.quota.ContainerQuota;
