+++ management/server/core/strategy-manager/strategy-manager-impl/src/main/java/io/subutai/core/strategy/impl/RoundRobinPlacementStrategy.java
@@ -13,7 +13,7 @@ import com.google.common.collect.Iterators;
 import io.subutai.common.environment.Node;
 import io.subutai.common.environment.NodeSchema;
 import io.subutai.common.environment.Topology;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.core.strategy.api.RoundRobinStrategy;
 import io.subutai.core.strategy.api.StrategyException;
 import io.subutai.hub.share.quota.ContainerQuota;
