+++ management/server/core/strategy-manager/strategy-manager-impl/src/main/java/io/subutai/core/strategy/impl/UnlimitedPlacementStrategy.java
@@ -11,7 +11,7 @@ import java.util.Set;
 import io.subutai.common.environment.Node;
 import io.subutai.common.environment.NodeSchema;
 import io.subutai.common.environment.Topology;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.core.strategy.api.StrategyException;
 import io.subutai.core.strategy.api.UnlimitedStrategy;
 import io.subutai.hub.share.quota.ContainerQuota;
