+++ management/server/subutai-common/src/main/java/io/subutai/common/environment/Node.java
@@ -7,7 +7,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 
 import io.subutai.common.gson.required.GsonRequired;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 
 
 /**
