+++ management/server/subutai-common/src/main/java/io/subutai/common/environment/NodeSchema.java
@@ -3,7 +3,7 @@ package io.subutai.common.environment;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 
 
 /**
