+++ management/server/subutai-common/src/main/java/io/subutai/common/peer/ContainerHost.java
@@ -6,6 +6,7 @@ import io.subutai.common.host.HostId;
 import io.subutai.common.metric.ProcessResourceUsage;
 import io.subutai.common.protocol.Template;
 import io.subutai.hub.share.quota.ContainerQuota;
+import io.subutai.hub.share.quota.ContainerSize;
 
 
 /**
