+++ management/server/subutai-common/src/main/java/io/subutai/common/peer/EnvironmentContainerHost.java
@@ -4,6 +4,7 @@ package io.subutai.common.peer;
 import java.util.Set;
 
 import io.subutai.common.security.SshKeys;
+import io.subutai.hub.share.quota.ContainerSize;
 
 
 public interface EnvironmentContainerHost extends ContainerHost
