+++ management/server/subutai-common/src/main/java/io/subutai/common/peer/Peer.java
@@ -36,7 +36,7 @@ import io.subutai.common.security.SshKeys;
 import io.subutai.common.security.relation.RelationLink;
 import io.subutai.common.security.relation.RelationLinkDto;
 import io.subutai.hub.share.quota.ContainerQuota;
-import io.subutai.hub.share.resource.ContainerResourceType;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.hub.share.resource.PeerResources;
