+++ management/server/subutai-common/src/main/java/io/subutai/common/peer/ResourceHost.java
@@ -16,6 +16,7 @@ import io.subutai.common.protocol.P2pIps;
 import io.subutai.common.protocol.Template;
 import io.subutai.common.protocol.Tunnel;
 import io.subutai.common.protocol.Tunnels;
+import io.subutai.hub.share.quota.ContainerSize;
 
 
 /**
