+++ management/server/subutai-common/src/main/java/io/subutai/common/task/CloneRequest.java
@@ -5,7 +5,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 
 import io.subutai.common.host.HostArchitecture;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 import io.subutai.common.settings.Common;
