+++ management/server/subutai-common/src/main/java/io/subutai/common/task/CloneResponse.java
@@ -5,7 +5,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.subutai.common.host.HostArchitecture;
-import io.subutai.common.peer.ContainerSize;
+import io.subutai.hub.share.quota.ContainerSize;
 
 
 public class CloneResponse implements TaskResponse
